Move state and functionality from the App component.
