+++ src/components/WatchList/WatchList.tsx
@@ -4,16 +4,16 @@ import { WatchCard } from '../WatchCard/WatchCard';
 import noWatch from '../../images/no-episode.gif';
 
 export const WatchList = ({
-    filteredEpisodes,
+    episodes,
     handleWatchList,
     handleRowClick
 }: {
-    filteredEpisodes: CleanEpisode[],
+    episodes: CleanEpisode[],
     handleWatchList: (id: number) => void,
     handleRowClick: (id: number) => void
 }) => {
 
-    const watchList = filteredEpisodes.filter(episode => episode.watchList)
+    const watchList = episodes.filter(episode => episode.watchList)
     const watchCards = watchList.map(episode => {
         return (
             <WatchCard
@@ -27,6 +27,7 @@ export const WatchList = ({
     return (
         <>
             <div className="container-watch-list">
+                <h3>My Watch List</h3>
                 {!watchCards.length ? <div className="container-message">
                     <p className="message-no-watch">Add episodes to your watch list to view them here!</p>
                     <img className="img-no-watch" src={noWatch} />
